Continued implementing Firestore interface
